refactor(uds): rehome validation and the option types in the model
Phase 2 of the format-blind extraction. §14.7 validation is model
semantics, so io/validate.rs becomes model/validate.rs; the option
vocabulary (FlowUnits, AnalysisOptions, LinkOffsets, Date and friends)
is what a run means, so it splits out of io/options.rs into
model/options.rs, leaving only the §14.4 grammar behind. Historical
io:: paths stay alive as re-export shims until the lift re-points
external consumers; every in-engine consumer is re-pointed now. The
audit this enables: io reaches the engine only through public items on
public module paths, so the physical lift cannot hit a visibility
wall.
